Split compare detail route sections
