+++ tests/compare-page-architecture.test.ts
@@ -24,6 +24,18 @@ const scorecardSource = readFileSync(
   'frontend/app/(localized)/[locale]/(marketing)/ai-video-engines/[slug]/_lib/compare-page-scorecard.ts',
   'utf8'
 );
+const copySource = readFileSync(
+  'frontend/app/(localized)/[locale]/(marketing)/ai-video-engines/[slug]/_lib/compare-page-copy.ts',
+  'utf8'
+);
+const specRowsSource = readFileSync(
+  'frontend/app/(localized)/[locale]/(marketing)/ai-video-engines/[slug]/_lib/compare-page-spec-rows.ts',
+  'utf8'
+);
+const detailContentSource = readFileSync(
+  'frontend/app/(localized)/[locale]/(marketing)/ai-video-engines/[slug]/_components/CompareDetailContent.tsx',
+  'utf8'
+);
 const generateCardSource = readFileSync(
   'frontend/app/(localized)/[locale]/(marketing)/ai-video-engines/[slug]/_components/CompareGenerateCard.tsx',
   'utf8'
@@ -32,19 +44,19 @@ const generateCardSource = readFileSync(
 test('comparison detail page delegates copy, helper, and media responsibilities', () => {
   const lineCount = pageSource.split('\n').length;
 
-  assert.ok(lineCount <= 1100, `comparison page should stay under 1100 lines after metadata split, got ${lineCount}`);
+  assert.ok(lineCount <= 500, `comparison page should stay under 500 lines after section extraction, got ${lineCount}`);
   assert.ok(pageSource.includes("from './_lib/compare-page-helpers'"));
   assert.ok(pageSource.includes("from './_lib/compare-page-overrides'"));
   assert.ok(pageSource.includes("from './_lib/compare-page-faq'"));
   assert.ok(pageSource.includes("from './_lib/compare-page-metadata'"));
   assert.ok(pageSource.includes("from './_lib/compare-page-scorecard'"));
-  assert.ok(pageSource.includes("from './_components/CompareGenerateCard'"));
-  assert.ok(pageSource.includes("from './_components/CompareShowdownMedia'"));
-  assert.ok(pageSource.includes("from './_components/CompareSpecValue'"));
+  assert.ok(pageSource.includes("from './_lib/compare-page-spec-rows'"));
+  assert.ok(pageSource.includes("from './_components/CompareDetailContent'"));
   assert.doesNotMatch(pageSource, /const COMPARE_PAGE_OVERRIDES/);
   assert.doesNotMatch(pageSource, /type ComparePageCopy =/);
   assert.doesNotMatch(pageSource, /const generatedFaqItems =/);
   assert.doesNotMatch(pageSource, /const comparisonMetrics = \[/);
+  assert.doesNotMatch(pageSource, /<section className=/);
   assert.doesNotMatch(pageSource, /buildSeoMetadata/);
 });
 
@@ -64,5 +76,10 @@ test('comparison detail split helpers own FAQ, scorecard, and generate card resp
   assert.match(scorecardSource, /export function deriveCompareStrengths/);
   assert.match(metadataSource, /export async function buildComparePageMetadata/);
   assert.match(metadataSource, /buildSeoMetadata/);
+  assert.match(copySource, /export function buildCompareDetailLabels/);
+  assert.match(specRowsSource, /export function buildCompareSpecRows/);
+  assert.match(detailContentSource, /export function CompareDetailContent/);
   assert.match(generateCardSource, /export function CompareGenerateCard/);
+  assert.match(detailContentSource, /CompareShowdownMedia/);
+  assert.match(detailContentSource, /CompareSpecValue/);
 });
